agents/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
neo4j>=5.0.0
prometheus-client>=0.20.0
boto3>=1.34.0
pyyaml>=6.0.0
structlog>=24.0.0
tenacity>=8.0.0